migrate: move healthCareLocalFacility to products directory
- Migrated healthCareLocalFacility product files to src/products/healthCareLocalFacility/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved all component files: Address, Phone, OperatingStatus, LocationServices, FacilitySocialLinks
- Moved HealthServices subdirectory with all component files
- Moved query.ts from src/data/queries/ to src/products/healthCareLocalFacility/
- Moved test file and renamed healthCareLocalFacility.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/healthCareLocalFacility/formatted-type.ts
- Moved both mock files: healthCareLocalFacility.mock.json -> mock.json and healthCareLocalFacility.mock.ts -> mock.ts
- Updated all import statements across the codebase including cross-references from other migrated products
- Fixed import paths in all migrated files
- Updated references in vamcSystem, vetCenter, vetCenterLocationListing, facilityListing, vbaFacility, locationsListing
- All TypeScript errors resolved
- All tests passing (16 query tests + 19 template tests + extensive component test coverage)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and healthCareLocalFacility pages load correctly (Pittsburgh facilities confirmed)
