+++ src/data/queries/index.ts
@@ -18,7 +18,7 @@ import * as EventTeaser from '@/data/queries/eventTeaser'
 import * as ExpandableText from '@/data/queries/expandableText'
 import * as FeaturedContent from '@/data/queries/featuredContent'
 import * as HeaderFooter from '@/data/queries/headerFooter'
-import * as HealthCareLocalFacility from '@/data/queries/healthCareLocalFacility'
+import * as HealthCareLocalFacility from '@/products/healthCareLocalFacility/query'
 import * as LinkTeaser from '@/data/queries/linkTeaser'
 import * as LocationsListing from '@/data/queries/locationsListing'
 import * as MediaDocument from '@/data/queries/mediaDocument'
