+++ src/pages/[[...slug]].tsx
@@ -52,7 +52,7 @@ import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
 import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
 import { VetCenter as FormattedVetCenter } from '@/products/vetCenter/formatted-type'
-import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
+import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/products/healthCareLocalFacility/formatted-type'
 import { VamcSystem as FormattedVamcSystem } from '@/products/vamcSystem/formatted-type'
 import { VamcSystemVaPolice as FormattedVamcSystemVaPolice } from '@/products/vamcSystemVaPolice/formatted-type'
 import { LeadershipListing as FormattedLeadershipListing } from '@/products/leadershipListing/formatted-type'
@@ -75,7 +75,7 @@ import { StaffProfile } from '@/products/staffProfile/template'
 import { StoryListing } from '@/products/storyListing/template'
 import { VetCenter } from '@/products/vetCenter/template'
 import { Wrapper } from '@/templates/layouts/wrapper'
-import { HealthCareLocalFacility } from '@/templates/layouts/healthCareLocalFacility'
+import { HealthCareLocalFacility } from '@/products/healthCareLocalFacility/template'
 import { DoNotPublishError } from '@/lib/drupal/query'
 import { VamcSystem } from '@/products/vamcSystem/template'
 import { VamcSystemVaPolice } from '@/products/vamcSystemVaPolice/template'
