+++ src/products/healthCareLocalFacility/HealthServices/HealthServices.tsx
@@ -1,4 +1,4 @@
-import { FormattedVAMCFacilityHealthService } from '@/types/formatted/healthCareLocalFacility'
+import { FormattedVAMCFacilityHealthService } from '@/products/healthCareLocalFacility/formatted-type'
 import { ServiceLocation } from './ServiceLocation'
 import { hashReference } from '@/lib/utils/hashReference'
 import { PhoneNumber } from '@/types/formatted/phoneNumber'
