+++ src/products/healthCareLocalFacility/mock.ts
@@ -5,7 +5,7 @@
  */
 
 import { NodeHealthCareLocalFacility } from '@/types/drupal/node'
-import deflatedData from './healthCareLocalFacility.mock.json'
+import deflatedData from './mock.json'
 import { inflateObjectGraph } from '@/lib/utils/object-graph'
 
 const mockData = inflateObjectGraph<NodeHealthCareLocalFacility>(deflatedData)
