+++ src/products/vamcSystem/formatted-type.ts
@@ -2,12 +2,12 @@ import { MediaImage } from '@/types/formatted/media'
 import { Administration } from '@/types/formatted/administration'
 import { SideNavMenu } from '@/types/formatted/sideNav'
 import { PublishedEntity } from '@/types/formatted/publishedEntity'
-import { HealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
+import { HealthCareLocalFacility } from '@/products/healthCareLocalFacility/formatted-type'
 import { FormattedRelatedLinks } from '@/types/formatted/relatedLinks'
 import { NodeHealthCareRegionPage } from '@/types/drupal/node'
 import { LovellChildVariant } from '@/lib/drupal/lovell/types'
 import { NewsStoryTeaser } from '@/products/newsStory/formatted-type'
-import { FacilitySocialLinksProps } from '@/templates/layouts/healthCareLocalFacility/FacilitySocialLinks'
+import { FacilitySocialLinksProps } from '@/products/healthCareLocalFacility/FacilitySocialLinks'
 import { EventWidgetTeaser } from '@/products/event/formatted-type'
 
 export type MinimalLocalFacility = Pick<
