+++ src/products/vamcSystem/template.tsx
@@ -11,7 +11,7 @@ import { LOVELL } from '@/lib/drupal/lovell/constants'
 import { ManageYourHealthLinks } from './ManageYourHealthLinks'
 import { LovellSwitcher } from '@/templates/components/lovellSwitcher'
 import { StoryTeaser } from './StoryTeaser'
-import FacilitySocialLinks from '@/templates/layouts/healthCareLocalFacility/FacilitySocialLinks'
+import FacilitySocialLinks from '@/products/healthCareLocalFacility/FacilitySocialLinks'
 // import { ListOfLinkTeasers } from '@/templates/components/listOfLinkTeasers'
 import { EventTeaser } from '@/templates/components/eventTeaser'
 // import { SocialLinks } from '@/templates/common/socialLinks'
