+++ src/products/vetCenter/template.tsx
@@ -8,7 +8,7 @@ import { QaSection } from '@/templates/components/qaSection'
 import { ExpandableOperatingStatus } from '@/templates/components/expandableOperatingStatus'
 import { PhoneNumber } from '@/templates/common/phoneNumber'
 import { SchemaScript } from './SchemaScript'
-import { Address } from '@/templates/layouts/healthCareLocalFacility/Address'
+import { Address } from '@/products/healthCareLocalFacility/Address'
 import { ContentFooter } from '@/templates/common/contentFooter'
 import { PrepareForVisitAccordions } from '@/templates/components/prepareForVisitAccordions'
 import { TextWithImage } from '@/templates/components/textWithImage'
