+++ src/templates/layouts/vbaFacility/index.tsx
@@ -3,7 +3,7 @@ import { ContentFooter } from '@/templates/common/contentFooter'
 import { VbaFacility as FormattedVBAFacility } from '@/types/formatted/vbaFacility'
 import { Wysiwyg } from '@/templates/components/wysiwyg'
 import { ExpandableOperatingStatus } from '@/templates/components/expandableOperatingStatus'
-import { Address } from '@/templates/layouts/healthCareLocalFacility/Address'
+import { Address } from '@/products/healthCareLocalFacility/Address'
 import { PhoneNumber } from '@/templates/common/phoneNumber'
 import { FeaturedContent } from '@/templates/common/featuredContent'
 import { Hours } from '@/templates/components/hours'
