+++ src/types/formatted/locationsListing.ts
@@ -2,7 +2,7 @@ import { PublishedEntity } from './publishedEntity'
 import { SideNavMenu } from '@/types/formatted/sideNav'
 import { NodeHealthCareRegionPage } from '../drupal/node'
 import { Administration } from '@/types/formatted/administration'
-import { HealthCareLocalFacility } from './healthCareLocalFacility'
+import { HealthCareLocalFacility } from '@/products/healthCareLocalFacility/formatted-type'
 import { LovellChildVariant } from '@/lib/drupal/lovell/types'
 
 export type MinimalLocalFacility = Pick<
